refactor: 将自定义无效参数异常类提取到shared域
