+++ cosmos-im-server/src/main/java/com/qfleaf/cosmosimserver/shared/exception/InvalidArgsException.java
@@ -1,0 +1,7 @@
+package com.qfleaf.cosmosimserver.shared.exception;
+
+public class InvalidArgsException extends BaseException {
+    public InvalidArgsException(String message) {
+        super(ErrorCode.INVALID_ARGS_ERROR, message);
+    }
+}
